Strategy, quote manager, order manager, portfolio, paper exchange, persistence
- strategy: Strategy trait + StrategyContext; basic_mm (fair ± half spread,
  inventory skew, levels, tick/step rounding, post-only clamp, fee check);
  fair_price local/weighted with renormalisation; inventory skew formula
- quote: QuoteManager requote thresholds + refresh throttle, forced on fills
- orders: state machine incl. CancelRequested->Filled race and Unknown;
  OrderManager desired-vs-live diff, in-flight guard, client id generator
- portfolio: holdings from balances, position from fills only, daily PnL
- exchange/paper: simulated account, locking, crossing fill model, tee of
  the wrapped market-data stream
- persistence: atomic JSON state store
- config: [paper] initial balances; MarketInfo rounding helpers

+++ src/exchange/binance/ws.rs
@@ -336,7 +336,7 @@ impl Backoff {
     pub fn next_delay(&mut self) -> Duration {
         let base = self.current;
         self.current = (self.current * 2).min(self.max);
-        let jitter_pct = (chrono::Utc::now().timestamp_subsec_nanos() % 26) as u32;
+        let jitter_pct = chrono::Utc::now().timestamp_subsec_nanos() % 26;
         base + base * jitter_pct / 100
     }
 }
